refactor(headless): single source of truth for heldInTaskSetHash (#1409)
Address review P2-a/P2-b/P2-c from the second independent pass.

P2-a — one source of truth. fixed-prompt-controller.ts duplicated
heldInTaskSetHash byte-for-byte (canonicalHeldInTaskSetHash) with no test
comparing the copies. Move heldInTaskSetHash + compareStrings + sortedUnique
into fixed-prompt-wal-types.ts (the shared upstream module both
fixed-prompt-controller and rsi-round-analysis already depend on), have
rsi-round-analysis re-export them, and drop both the
canonicalHeldInTaskSetHash duplicate and the now-unused hashHeldInTaskSet
alias from prompt-candidate-loop.ts.

P2-b — smoke test was tautological. The cross-WAL-boundary test set both
sides via the same alias; import heldInTaskSetHash directly and refresh the
stale "formerly localeCompare" comment to reflect that both sides now use the
unified codepoint hash (writer-side regression guard lives in the
writer-contract test).

P2-c — document the projection's tamper semantics. readFixedPromptWal
re-derives the committed hash on every read, so hash-only tampering is
silently normalized before the self-consistency check (ids tampering is still
caught); add a comment spelling out this trade-off and an Array.isArray guard
on heldInTaskIds.
